Shuffle some test initialization code to allow running under xdist.
