+++ pytest.ini
@@ -3,4 +3,8 @@ addopts = --reuse-db --tb=native
 python_files=test*.py
 markers =
     es_tests: mark a test as an elasticsearch test.
-norecursedirs = *.egg venv src .tox
+norecursedirs =
+    node_modules locale static media site-static user-media tmp
+    templates fixtures migrations
+    .* *.egg src dist cache venv __pycache__
+DJANGO_SETTINGS_MODULE = settings_test
